Fix a little annoyance with hot module reloading
Hey,

Thank you for this package! I'm using it with react-leaflet for a remix-run app.

I consistently have an issue when editing code, and hot module reload happens. It seems that somehow the map instance gets destroyed before the _unload() event handler runs, hence this patch.

Let me know if you think there's a better way to do it.